Inner classes with builder don't get introspected (#11313)
When an inner class is postponed to the next round the binary class name is used to perform a lookup of the refreshed class when processing gets to the next round. This is an issue because the canonical class name needs to be use to lookup elements. This changes the code to use `getCanonicalName()` which revealed another bug where the implementation of `JavaClassElement.getCanonicalName()` was not correct.

---------

Co-authored-by: Graeme Rocher <graeme.rocher@oracle.com>