Update MicronautMetaServiceLoaderUtils.java to fix single threaded service loader #11229  (#11326)
Fix for service loader #11229

I believe the change is not complete and still has a logical error. Another way to change the parallelisim would be to set -XX:ActiveProcessorCount. This option is used in dockerized environments to limit resources used by the application.

In my tests, wenn I set the option a value smaller than 3 (parallelism is active processor count -1), then micronaut application will not start.

I believe the condition to add the services in the class MicronautMetaServiceLoaderUtils (line 260) is not correct:

is:

```
if (val != null && predicate != null && !predicate.test(val)) {
```

but should be:

```
if (val != null && !collection.contains(val) && (predicate == null || predicate.test(val))) {
```